mine/gemini20-all/requirements.txt
google-genai
numpy
SpeechRecognition
websockets
markdown
pygments
pypdf
python-dotenv
chromadb